Changed RGB data type to integer.

+++ expense-tracker-server/src/main/java/inzagher/expense/tracker/server/dto/ColorDTO.java
@@ -7,40 +7,40 @@ import javax.xml.bind.annotation.XmlType;
 @XmlType
 @XmlAccessorType(XmlAccessType.FIELD)
 public class ColorDTO {
-    private byte red;
-    private byte green;
-    private byte blue;
+    private Integer red;
+    private Integer green;
+    private Integer blue;
     
     public ColorDTO() {
     }
     
-    public ColorDTO(byte red, byte green, byte blue) {
+    public ColorDTO(Integer red, Integer green, Integer blue) {
         this.red = red;
         this.green = green;
         this.blue = blue;
     }
 
-    public byte getRed() {
+    public Integer getRed() {
         return red;
     }
 
-    public void setRed(byte red) {
+    public void setRed(Integer red) {
         this.red = red;
     }
 
-    public byte getGreen() {
+    public Integer getGreen() {
         return green;
     }
 
-    public void setGreen(byte green) {
+    public void setGreen(Integer green) {
         this.green = green;
     }
 
-    public byte getBlue() {
+    public Integer getBlue() {
         return blue;
     }
 
-    public void setBlue(byte blue) {
+    public void setBlue(Integer blue) {
         this.blue = blue;
     }
 }
