+++ expense-tracker-server/src/test/java/inzagher/expense/tracker/server/CategoryServiceTests.java
@@ -34,8 +34,8 @@ public class CategoryServiceTests {
     @BeforeEach
     public void beforeEachTest() {
         eric = manager.storePerson("ERIC");
-        phone = manager.storeCategory("PHONE", "MONTHLY PHONE BILL", (byte)0, (byte)0, (byte)0);
-        rent = manager.storeCategory("RENT", "MONTHLY RENT BILL", (byte)128, (byte)128, (byte)128);
+        phone = manager.storeCategory("PHONE", "MONTHLY PHONE BILL", 0, 0, 0);
+        rent = manager.storeCategory("RENT", "MONTHLY RENT BILL", 128, 128, 128);
         payment = manager.storeExpense(2020, 10, 10, eric, phone, 1000.50D, "TEST BILL PAYMENT");
         assertEquals(1L, manager.countPersons());
         assertEquals(2L, manager.countCategories());
@@ -69,7 +69,7 @@ public class CategoryServiceTests {
         CategoryDTO education = new CategoryDTO();
         education.setName("EDUCATION");
         education.setDescription("YEARLY EDUCATION BILL");
-        education.setColor(new ColorDTO((byte)30, (byte)30, (byte)30));
+        education.setColor(new ColorDTO(30, 30, 30));
         UUID storedRecordID = categoryService.storeCategory(education);
         assertNotNull(storedRecordID);
         assertEquals(3L, manager.countCategories());
@@ -79,7 +79,7 @@ public class CategoryServiceTests {
     @Test
     public void categoryEditingTest() {
         CategoryDTO category = phone.toDTO();
-        category.setColor(new ColorDTO((byte)16, (byte)16, (byte)16));
+        category.setColor(new ColorDTO(16, 16, 16));
         UUID storedRecordID = categoryService.storeCategory(category);
         assertEquals(phone.getId(), storedRecordID);
         assertEquals(2L, manager.countCategories());
